use `witx docs --check` to replace specialized tests
