+++ tools/witx/cli/src/main.rs
@@ -4,7 +4,7 @@ use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::process;
 use structopt::{clap::AppSettings, StructOpt};
-use witx::{load, phases, Document, Documentation};
+use witx::{load, Document, Documentation};
 
 /// Validate and process witx files
 #[derive(StructOpt, Debug)]
@@ -31,6 +31,9 @@ enum Command {
         /// Path to root of witx document
         #[structopt(number_of_values = 1, value_name = "INPUT", parse(from_os_str))]
         input: Vec<PathBuf>,
+        /// Perform check that output matches witx documents
+        #[structopt(long = "check")]
+        check: bool,
         /// Path to generated documentation in Markdown format
         #[structopt(
             short = "o",
@@ -40,8 +43,6 @@ enum Command {
         )]
         output: Option<PathBuf>,
     },
-    /// Update documentation in WASI repository to reflect witx specs
-    RepoDocs,
     /// Examine differences between interfaces
     Polyfill {
         /// Path to root of witx document
@@ -80,22 +81,32 @@ pub fn main() {
     let verbose = args.verbose;
 
     match args.cmd {
-        Command::Docs { input, output } => {
+        Command::Docs {
+            input,
+            check,
+            output,
+        } => {
             let doc = load_witx(&input, "input", verbose);
-            if let Some(output) = output {
-                write_docs(&doc, output)
+            if check {
+                let output = output.expect("output argument required in docs --check mode");
+                if diff_against_filesystem(&doc.to_md(), &output).is_err() {
+                    println!("Docs in tree are out-of-date with witx files. Re-run this executable with the following arguments to to re-generate:");
+                    println!(
+                        "docs {} --output {}",
+                        input
+                            .iter()
+                            .map(|p| p.to_string_lossy().into_owned())
+                            .collect::<Vec<String>>()
+                            .join(" "),
+                        output.to_string_lossy(),
+                    );
+                }
             } else {
-                println!("{}", doc.to_md())
-            }
-        }
-        Command::RepoDocs => {
-            for phase in &[
-                phases::snapshot().unwrap(),
-                phases::ephemeral().unwrap(),
-                phases::old::snapshot_0().unwrap(),
-            ] {
-                let doc = load(&phase).expect("parse phase");
-                write_docs(&doc, phases::docs_path(&phase));
+                if let Some(output) = output {
+                    write_docs(&doc, output)
+                } else {
+                    println!("{}", doc.to_md())
+                }
             }
         }
         Command::Polyfill {
@@ -172,4 +183,88 @@ fn parse_module_mapping(m: &str) -> Result<(String, String)> {
         _ => bail!("invalid module mapping: '{}'", m),
     };
     Ok((n.to_string(), o.to_string()))
+}
+
+fn dos2unix(s: &str) -> String {
+    let mut t = String::new();
+    t.reserve(s.len());
+    for c in s.chars() {
+        if c != '\r' {
+            t.push(c)
+        }
+    }
+    t
+}
+
+fn diff_against_filesystem(expected: &str, path: &Path) -> Result<(), ()> {
+    let actual = std::fs::read_to_string(path)
+        .unwrap_or_else(|e| panic!("couldn't read {}: {:?}", Path::display(path), e));
+    // Git may checkout the file with dos line endings on windows. Strip all \r:
+    let actual = dos2unix(&actual);
+    if &actual == expected {
+        return Ok(());
+    }
+
+    eprintln!("The following diff was found between the docs generated from .witx and the");
+    eprintln!("source {:?} in the tree:", path);
+    eprintln!();
+
+    let mut expected_line = 1;
+    let mut actual_line = 1;
+    let mut separated = false;
+    let mut any_lines = false;
+    for diff in diff::lines(&expected, &actual) {
+        match diff {
+            diff::Result::Left(l) => {
+                eprintln!("line {}: -{}", expected_line, l);
+                expected_line += 1;
+                separated = false;
+                any_lines = true;
+            }
+            diff::Result::Both(_, _) => {
+                expected_line += 1;
+                actual_line += 1;
+                if !separated {
+                    eprintln!("...");
+                    separated = true;
+                }
+            }
+            diff::Result::Right(r) => {
+                eprintln!("line {}: +{}", actual_line, r);
+                actual_line += 1;
+                separated = false;
+                any_lines = true;
+            }
+        }
+    }
+
+    if !any_lines {
+        eprintln!();
+        eprintln!(
+            "Somehow there was a diff with no lines differing. Lengths: {} and {}.",
+            expected.len(),
+            actual.len()
+        );
+        for (index, (a, b)) in actual.chars().zip(expected.chars()).enumerate() {
+            if a != b {
+                eprintln!("char difference at index {}: '{}' != '{}'", index, a, b);
+            }
+        }
+        for (index, (a, b)) in actual.bytes().zip(expected.bytes()).enumerate() {
+            if a != b {
+                eprintln!("byte difference at index {}: b'{}' != b'{}'", index, a, b);
+            }
+        }
+        eprintln!();
+        eprintln!("actual: {}", actual);
+        eprintln!();
+        eprintln!("expected: {}", expected);
+    }
+
+    eprintln!();
+    eprintln!(
+        "To regenerate the files, run `cd tools/witx && cargo run --example witx repo-docs`."
+    );
+    eprintln!();
+    Err(())
 }
